# Conflicts:
#	szcshl-web/src/main/java/cs/domain/financial/FinancialManager.java
#	szcshl-web/src/main/java/cs/service/financial/FinancialManagerService.java
#	szcshl-web/src/main/java/cs/service/financial/FinancialManagerServiceImpl.java
#	szcshl-web/src/main/webapp/WEB-INF/views/financial/assistCostList.html
#	szcshl-web/src/main/webapp/contents/app/financial/assistCostCount.edit.js

+++ szcshl-web/src/main/java/cs/service/financial/FinancialManagerService.java
@@ -1,5 +1,6 @@
 package cs.service.financial;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
@@ -25,7 +26,8 @@ public interface FinancialManagerService {
 
 	void delete(String id);
 
-	Integer sunCount(String signid);
+
+	BigDecimal sunCount(String signid);
 
 	Map<String, Object> initfinancialData(String signid);
